Remove unnecessary class FSM
